+++ src/main/kotlin/com/atlassian/performance/tools/jiraactions/api/page/ProjectSummaryPage.kt
@@ -2,18 +2,14 @@ package com.atlassian.performance.tools.jiraactions.api.page
 
 import org.openqa.selenium.By
 import org.openqa.selenium.WebDriver
-import org.openqa.selenium.support.ui.ExpectedConditions
-import java.time.Duration
+import org.openqa.selenium.support.ui.ExpectedConditions.presenceOfElementLocated
 
 class ProjectSummaryPage(
     private val driver: WebDriver
 ) {
 
     fun waitForMetadata(): ProjectSummaryPage {
-        driver.wait(
-            Duration.ofSeconds(6),
-            ExpectedConditions.presenceOfElementLocated(By.className("project-meta-column"))
-        )
+        driver.wait(presenceOfElementLocated(By.className("project-meta-column")))
         return this
     }
 }